Fix pt-BR locale code inside files
I asked for support to find the reason of this kind of commits [1] from
Crowdin integration, they fixed the problem by configuring an internal
post-processor for this project so the reason seems to be some change on
the Crowdin side. I tested it successfully by making a dummy update on one
translation to see if next commits for this language will come with the
right locale, as you can see at following commit [2] it worked!

[1] https://github.com/consul/consul/pull/3944/commits/6f1bb89ec3d582885c8ed8c47b736e68016364e7
[2] https://github.com/consul/consul/pull/3944/commits/2f3ed05477c27bc0b487aff08ff21292b7f9a222