Enable Serbian (Cyrillic)
Consul uses two variants of Serbian: Serbian (Cirylic), and
Serbo-Croatian so wee need to use different locale codes for them.

The folder names of both languages were correct but the locale code within
the files was the same for both languages "sr" so translations were mixed.

This commit changes the locale code for Serbian (Cyrillic) translations
files to "sr-SP", this will be done by Crowdin automatically with a
new postprocessor Crowdin configures internally so we will not have to
update the locale code for this language anymore.